Ignore Conversion Comments
The current behavior of smartconnector.py will add the conversion comments from SmartMove into custom-fields.field-1. 
This field has a length limitation of 254 characters and will cause rules to not import when running SmartConnector as the field length is not currently checked or truncated. 
This information is also in the conversion report, so some users might not want this data imported into their policies. 

This provides a --ignore-conversion-comments flag to simply skip adding them to the payload.  
There is a similar concern with the bash scripts that are generated, but this does not resolve that issue.